+++ src/drivers/pm1006.h
@@ -1,7 +1,3 @@
-#include <Arduino.h>
-
-#include <stdbool.h>
-#include <stdint.h>
 
 typedef enum {
     PM1006_HEADER,
